requirements.txt
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
